refactor: move IncludeValidationPipe to common module
