+++ src/common/pipes/include-validation.pipe.ts
@@ -1,12 +1,19 @@
 import { PipeTransform, Injectable } from "@nestjs/common";
 import { BadRequestException } from "@nestjs/common/exceptions";
+import { PipelineStage } from "mongoose";
 import { isJsonString } from "src/common/utils";
-import { DATASET_LOOKUP_FIELDS } from "src/datasets/types/dataset-lookup";
+import { DatasetLookupKeysEnum } from "src/datasets/types/dataset-lookup";
+import { OrigDatablockLookupKeysEnum } from "src/origdatablocks/types/origdatablock-lookup";
 
 @Injectable()
 export class IncludeValidationPipe
   implements PipeTransform<string | string[], string | string[]>
 {
+  constructor(
+    private lookupFields:
+      | Record<DatasetLookupKeysEnum, PipelineStage.Lookup | undefined>
+      | Record<OrigDatablockLookupKeysEnum, PipelineStage.Lookup | undefined>,
+  ) {}
   transform(inValue: string | string[]): string[] | string {
     if (!inValue) {
       return inValue;
@@ -20,7 +27,7 @@ export class IncludeValidationPipe
         : Array(inValue);
 
     includeValueParsed?.map((field) => {
-      if (Object.keys(DATASET_LOOKUP_FIELDS).includes(field)) {
+      if (Object.keys(this.lookupFields).includes(field)) {
         return field;
       } else {
         throw new BadRequestException(
